requirements.txt
pyrogram>=2.0.106
tgcrypto>=1.2.5
pymongo>=4.3.3
GitPython>=3.1.31
environs>=9.5.0
pillow>=10.3.0
requests>=2.28.2
aiohttp>=3.8.4
aiofiles>=23.1.0
beautifulsoup4>=4.12.2
humanize>=4.6.0
pySmartDL>=1.3.4
lexica-api>=0.2.0
dnspython>=2.3.0
psutil>=5.9.5